+++ sickrage/core/__init__.py
@@ -94,6 +94,7 @@ class Core(object):
         self.newest_version = None
         self.newest_version_string = None
 
+        self.api_url = 'https://api.sickrage.ca'
         self.user_agent = 'SiCKRAGE.CE.1/({};{};{})'.format(platform.system(), platform.release(), str(uuid.uuid1()))
         self.sys_encoding = get_sys_encoding()
         self.languages = [language for language in os.listdir(sickrage.LOCALE_DIR) if '_' in language]